add project

+++ src/main/java/com/example/optika/controller/BuyOrderController.java
@@ -7,13 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.util.Map;
 
 @Controller
 public class BuyOrderController {
@@ -28,11 +23,8 @@ public class BuyOrderController {
         return "/order/orders";
     }
 
-    //сделать отдельную страницу ордеров для админа
     @GetMapping("/thx")
     public String thxPage(Model model) {
-//        Iterable<BuyOrder> order = buyOrderRepository.findAll();
-//        model.addAttribute("orders", order);
         return "/order/thx";
     }
 
@@ -43,9 +35,8 @@ public class BuyOrderController {
         return "/order/order-list";
     }
 
-//    @PathVariable("id") long id,
     @PostMapping("/orders/{id}")
-    public String addProduct(@RequestParam int phoneNumber, @RequestParam String message, Product product)  {
+    public String addProduct(@RequestParam Long phoneNumber, @RequestParam String message, Product product)  {
 
         BuyOrder buyOrder = new BuyOrder(phoneNumber, message, product);
         buyOrderRepository.save(buyOrder);
@@ -53,16 +44,10 @@ public class BuyOrderController {
         return "redirect:/thx";
     }
 
-
-    // при нажатаии Buy отобразить товар, № тел, сообщение + редірекст на спс страницу
-    // создать страницу ордеров для админа
-    //перевести БД с инт на лонг
-
     @GetMapping("/deleteOrder")
     public String deleteBueOrder(@RequestParam Long orderId) {
         buyOrderRepository.deleteById(orderId);
         return "redirect:/";
     }
-
 
 }
